requirements.txt
# Framework Web
fastapi==0.120.2
uvicorn==0.38.0

# Validación de datos
pydantic==2.12.3

# HTTP Client
requests==2.32.5

# SDK de Transbank para Webpay Plus
transbank-sdk==6.1.0

# Herramientas de desarrollo (opcional)
pytest==8.3.3
pytest-cov==6.0.0
httpx==0.27.2
